requirements.txt
json
langchain
chat
openai
streamlit
OpenAI
os
pandas
rag
PIL
chat
numpy
Image
load_chain
initialise_prompt
prompts
pytesseract
pdf2image
PyPDF2
langchain_openai
PdfReader
convert_from_bytes
PdfReader
langchain_openai  
ChatOpenAI
langchain.chains
ConversationChain
langchain.memory
ConversationBufferMemory
docx
ai_prompt
ai_topic_prompt
latex_prompt
functions
read_word_table
create_word_doc
get_text
subprocess
re
dotenv
load_dotenv
docxlatex
Document
base64
fixed_function
io
BytesIO
sys
langchain.vectorstores
Chroma
langchain_openai
ChatOpenAI
langchain.text_splitter
CharacterTextSplitter
langchain.llms
langchain.chains
VectorDBQA
langchain.document_loaders 
DirectoryLoader
langchain.text_splitter 
RecursiveCharacterTextSplitter
langchain.embeddings 
HuggingFaceInstructEmbeddings
langchain.chains
ConversationalRetrievalChain
ConversationChain
langchain.memory 
ChatMessageHistory
ConversationBufferMemory
langchain.embeddings.azure_openai
AzureOpenAIEmbeddings
langchain.document_loaders
PyPDFLoader
langchain.document_loaders 
Docx2txtLoader
langchain.document_loaders 
TextLoader
langchain.document_loaders 
UnstructuredMarkdownLoader
langchain.document_loaders 
UnstructuredPowerPointLoader
langchain.prompts 
PromptTemplate
langchain.chains.question_answering 
load_qa_chain
langchain.chat_models 
AzureChatOpenAI
pytesseract
Output
TesseractError
functions 
convert_pdf_to_txt_pages 
convert_pdf_to_txt_file 
save_pages 
displayPDF 
images_to_txt
filter_prompt2,
initialise_prompt
master_prompt
lang_prompt
format_prompt
ai_prompt
aiformat_prompt
mcq_test_prompt
key_term_prompt
learn_outcome_prompt
student_prompt
chromadb
nltk
constants
docx
tempfile
re
typing 
Union
List
weaviate
db
utils
pandas
docx2txt
docxlatex
Document 
zipfile 
pdfminer.pdfinterp 
PDFResourceManager
PDFPageInterpreter
pdfminer.converter 
TextConverter
pdfminer.layout 
LAParams
pdfminer.pdfpage
PDFPage
StringIO
Output
easyocr
cv2
numpy
logging
typing
Callable